capabilities,general: move RegularExpressionsClientCapabilities
